Move `Parameters` to dedicated module
